readability + reusability refactor
