+++ apps/client/types.d.ts
@@ -1,0 +1,6 @@
+export interface Trade {
+  ticker: string;
+  quantity: number;
+  price: number;
+  type: string;
+}
